细节修正
+ 404 页面
+ select 选择文件名格式
+ 登录规则，添加redis缓存和cookie双认证记录登录状态，用session保存用户信息

+++ src/main/java/kid1999/upload/controller/download.java
@@ -1,5 +1,6 @@
 package kid1999.upload.controller;
 
+import com.sun.org.apache.xpath.internal.operations.Mod;
 import kid1999.upload.service.homeworkService;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -34,37 +36,40 @@ public class download {
 	// 下载文件
 	@PostMapping("download")
 	ResponseEntity<byte[]> download(HttpServletRequest request,
-	                           Model model) throws Exception {
+									HttpServletResponse response) throws Exception {
 		String userid = request.getParameter("studentid");
-		String[] filenames = request.getParameterValues("filenames");
+		if(request.getParameterValues("filenames") == null){
+			response.sendRedirect(request.getHeader("REFERER"));
+			return null;
+		}else{
+			String[] filenames = request.getParameterValues("filenames");
+			int uid = Integer.parseInt(userid);
+			String addr = homeworkService.findaddrBySid(uid);
+			String tmpfile = Paths.get(TMPFILEPATH,userid).toString() + ".zip";    //缓存文件
 
-		int uid = Integer.parseInt(userid);
-		String addr = homeworkService.findaddrBySid(uid);
-		String tmpfile = Paths.get(TMPFILEPATH,userid).toString() + ".zip";    //缓存文件
+			ZipOutputStream zipOut = new ZipOutputStream(new FileOutputStream(tmpfile));  // 输出文件
+			InputStream input = null;
 
-		ZipOutputStream zipOut = new ZipOutputStream(new FileOutputStream(tmpfile));  // 输出文件
-		InputStream input = null;
-
-		// 将选择文件一一写入
-		for (String str : filenames) {
-			input = new FileInputStream(new File(addr,str));
-			zipOut.putNextEntry(new ZipEntry(str));
-			byte[] buff = new byte[1024];
-			while((input.read(buff)) != -1){
-				zipOut.write(buff);
+			// 将选择文件一一写入
+			for (String str : filenames) {
+				input = new FileInputStream(new File(addr,str));
+				zipOut.putNextEntry(new ZipEntry(str));
+				byte[] buff = new byte[1024];
+				while((input.read(buff)) != -1){
+					zipOut.write(buff);
+				}
+				input.close();
 			}
-			input.close();
-		}
-		zipOut.close();
-		File file = new File(tmpfile);
-		HttpHeaders headers = new HttpHeaders();
+			zipOut.close();
+			File file = new File(tmpfile);
+			HttpHeaders headers = new HttpHeaders();
 //		String filename = new String(tmpfile.getBytes("utf-8"),"iso-8859-1");   //中文文件名
-		headers.setContentDispositionFormData("attachment", "result.zip");  //设置下载文件名
-		headers.setContentType(MediaType.APPLICATION_OCTET_STREAM);
-		ResponseEntity<byte[]> response = new ResponseEntity<byte[]>(FileUtils.readFileToByteArray(file),headers, HttpStatus.CREATED);
-		Files.deleteIfExists(Paths.get(tmpfile));   // 清除缓存
-		return response;
-
+			headers.setContentDispositionFormData("attachment", "result.zip");  //设置下载文件名
+			headers.setContentType(MediaType.APPLICATION_OCTET_STREAM);
+			ResponseEntity<byte[]> res = new ResponseEntity<byte[]>(FileUtils.readFileToByteArray(file),headers, HttpStatus.CREATED);
+			Files.deleteIfExists(Paths.get(tmpfile));   // 清除缓存
+			return res;
+		}
 	}
 
 }
